0209 B4

+++ src/B5/Day15/BOJ11382.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class BOJ11382 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int A = scanner.nextInt();
-        int B = scanner.nextInt();
-        int C = scanner.nextInt();
+        long A = scanner.nextLong();
+        long B = scanner.nextLong();
+        long C = scanner.nextLong();
 
         System.out.println(A+B+C);
     }
